backend/requirements.txt
colorama==0.4.6
flask==3.1.0
flask-cors==5.0.1
joblib==1.4.2
numpy==2.2.4
python-dotenv==1.1.0
scikit-learn==1.6.1
scipy==1.15.2
werkzeug==3.1.3
xgboost==3.0.0
pandas==2.2.3
requests
boto3
firebase-admin
